+++ src/sass.rs
@@ -13,6 +13,10 @@ pub enum SassError {
 /// Render a SASS/SCSS file to the destination.
 /// Other SASS/SCSS files next to the provided one will be available for
 /// inclusion.
+///
+/// # Errors
+///
+/// Returns an error if the SASS file cannot be compiled or if the output cannot be written.
 pub fn render_sass(source: impl AsRef<Path>, path: impl AsRef<Path>) -> Result<String, SassError> {
     let source = source.as_ref();
     let options = match source.parent() {